Innfør TypeScript på server, refaktor
