+++ .github/workflows/deploy.yml
@@ -24,6 +24,10 @@ jobs:
               run: npm run test
             - name: Bygg applikasjon
               run: npm run build
+            - name: Installer avhengigheter for server
+              run: cd server && npm ci
+            - name: Bygg server
+              run: cd server && npm run build
             - name: Opprett release med Sentry
               run: npm run sentry-release
               env:
