+++ server/package.json
@@ -1,5 +1,16 @@
 {
+    "scripts": {
+        "build": "tsc --build",
+        "start": "nodemon src/server.ts"
+    },
     "dependencies": {
         "http-proxy-middleware": "^1.0.6"
+    },
+    "devDependencies": {
+        "@types/express": "^4.17.13",
+        "@types/node": "^17.0.18",
+        "nodemon": "^2.0.15",
+        "ts-node": "^10.5.0",
+        "typescript": "^4.5.5"
     }
 }
