+++ server/tsconfig.json
@@ -1,0 +1,11 @@
+{
+    "compilerOptions": {
+        "noEmitOnError": true,
+        "removeComments": false,
+        "target": "es5",
+        "esModuleInterop": true,
+        "skipLibCheck": true,
+        "outDir": "build"
+    },
+    "include": ["src/**/*"]
+}
